+++ test/backgroundScripts/autoRedirect.test.js
@@ -1,0 +1,1 @@
+// Nothing here until it's a defined feature
